refactor(frontend): use repositories index import
