+++ frontend/__tests__/.server/domain/repositories/marital-status.repository.test.ts
@@ -1,7 +1,7 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 import { mock } from 'vitest-mock-extended';
 
-import { MaritalStatusRepositoryImpl } from '~/.server/domain/repositories/marital-status.repository';
+import { MaritalStatusRepositoryImpl } from '~/.server/domain/repositories';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
 
 const dataSource = vi.hoisted(() => ({
