+++ frontend/__tests__/.server/domain/repositories/preferred-communication-method.repository.test.ts
@@ -1,7 +1,7 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 import { mock } from 'vitest-mock-extended';
 
-import { PreferredCommunicationMethodRepositoryImpl } from '~/.server/domain/repositories/preferred-communication-method.repository';
+import { PreferredCommunicationMethodRepositoryImpl } from '~/.server/domain/repositories';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
 
 const dataSource = vi.hoisted(() => ({
