+++ frontend/__tests__/.server/domain/services/marital-status.service.test.ts
@@ -4,7 +4,7 @@ import { mock } from 'vitest-mock-extended';
 import type { ServerConfig } from '~/.server/configs/server.config';
 import type { MaritalStatusDto } from '~/.server/domain/dtos';
 import type { MaritalStatusDtoMapper } from '~/.server/domain/mappers';
-import type { MaritalStatusRepository } from '~/.server/domain/repositories/marital-status.repository';
+import type { MaritalStatusRepository } from '~/.server/domain/repositories';
 import { MaritalStatusServiceImpl } from '~/.server/domain/services';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
