+++ frontend/__tests__/.server/domain/services/preferred-communication-method.service.test.ts
@@ -4,7 +4,7 @@ import { mock } from 'vitest-mock-extended';
 import type { ServerConfig } from '~/.server/configs/server.config';
 import type { PreferredCommunicationMethodDto } from '~/.server/domain/dtos';
 import type { PreferredCommunicationMethodDtoMapper } from '~/.server/domain/mappers';
-import type { PreferredCommunicationMethodRepository } from '~/.server/domain/repositories/preferred-communication-method.repository';
+import type { PreferredCommunicationMethodRepository } from '~/.server/domain/repositories';
 import { PreferredCommunicationMethodServiceImpl } from '~/.server/domain/services';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
