+++ frontend/app/.server/domain/services/client-application.service.ts
@@ -3,7 +3,7 @@ import { inject, injectable } from 'inversify';
 import { SERVICE_IDENTIFIER } from '~/.server/constants/service-identifier.contant';
 import type { ClientApplicationDto } from '~/.server/domain/dtos';
 import type { ClientApplicationDtoMapper } from '~/.server/domain/mappers';
-import type { ClientApplicationRepository } from '~/.server/domain/repositories/client-application.repository';
+import type { ClientApplicationRepository } from '~/.server/domain/repositories';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
 
 export interface FindByFirstNameLastNameDobClientNumberSearchCriteria {
