+++ frontend/app/.server/domain/services/client-friendly-status.service.ts
@@ -5,7 +5,7 @@ import type { ServerConfig } from '~/.server/configs/server.config';
 import { SERVICE_IDENTIFIER } from '~/.server/constants/service-identifier.contant';
 import type { ClientFriendlyStatusDto } from '~/.server/domain/dtos';
 import type { ClientFriendlyStatusDtoMapper } from '~/.server/domain/mappers';
-import type { ClientFriendlyStatusRepository } from '~/.server/domain/repositories/client-friendly-status.repository';
+import type { ClientFriendlyStatusRepository } from '~/.server/domain/repositories';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
 
 export interface ClientFriendlyStatusService {
