+++ frontend/app/.server/domain/services/provincial-government-insurance-plan.service.ts
@@ -5,7 +5,7 @@ import type { ServerConfig } from '~/.server/configs/server.config';
 import { SERVICE_IDENTIFIER } from '~/.server/constants/service-identifier.contant';
 import type { ProvincialGovernmentInsurancePlanDto } from '~/.server/domain/dtos';
 import type { ProvincialGovernmentInsurancePlanDtoMapper } from '~/.server/domain/mappers';
-import type { ProvincialGovernmentInsurancePlanRepository } from '~/.server/domain/repositories/provincial-government-insurance-plan.repository';
+import type { ProvincialGovernmentInsurancePlanRepository } from '~/.server/domain/repositories';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
 
 export interface ProvincialGovernmentInsurancePlanService {
